+++ src/datasources/alarms/query-type-handlers/AlarmsQueryHandlersCore.test.ts
@@ -4,7 +4,30 @@ import { AlarmsQuery, QueryAlarmsRequest, QueryType } from '../types/types';
 import { MockProxy } from 'jest-mock-extended';
 import { BackendSrv } from '@grafana/runtime';
 import { createFetchError, createFetchResponse, requestMatching, setupDataSource } from 'test/fixtures';
-import { QUERY_ALARMS_RELATIVE_PATH } from '../constants/QueryAlarms.constants';
+import { QUERY_ALARMS_RELATIVE_PATH } from './constants/QueryAlarms.constants';
+import { Workspace } from 'core/types';
+
+jest.mock('core/utils', () => ({
+  getVariableOptions: jest.fn(),
+}));
+
+jest.mock('shared/workspace.utils', () => {
+  return {
+    WorkspaceUtils: jest.fn().mockImplementation(() => ({
+      getWorkspaces: jest.fn().mockResolvedValue(
+        new Map([
+          ['Workspace1', { id: 'Workspace1', name: 'Workspace Name' }],
+          ['Workspace2', { id: 'Workspace2', name: 'Another Workspace Name' }],
+        ])
+      )
+    }))
+  };
+});
+import { getVariableOptions } from 'core/utils';
+
+jest.mock('core/utils', () => ({
+  getVariableOptions: jest.fn(),
+}));
 
 class TestAlarmsDataSource extends AlarmsQueryHandlersCore {
   async runQuery(query: AlarmsQuery, _: DataQueryRequest): Promise<DataFrameDTO> {
@@ -33,6 +56,21 @@ describe('AlarmsQueryHandlersCore', () => {
 
   beforeEach(() => {
     [datastore, backendServer] = setupDataSource(TestAlarmsDataSource);
+  });
+
+  describe('globalVariableOptions', () => {
+    it('should get variable options', () => {
+      const mockOptions = [
+        { label: 'Variable 1', value: '$var1' },
+        { label: 'Variable 2', value: '$var2' },
+      ];
+      (getVariableOptions as jest.Mock).mockReturnValue(mockOptions);
+
+      const result = datastore.globalVariableOptions();
+
+      expect(getVariableOptions).toHaveBeenCalledWith(datastore);
+      expect(result).toEqual(mockOptions);
+    });
   });
 
   describe('queryAlarms', () => {
@@ -153,6 +191,113 @@ describe('AlarmsQueryHandlersCore', () => {
         expect(datastore.templateSrv.replace).toHaveBeenCalledWith('alarmId < "${query0}"', {});
         expect(transformQuery).toBe('alarmId < "test-alarmID-1"');
       });
+
+      it('should replace multiple value variable in the filter', () => {
+        const mockFilter = 'channel = "${query0}"';
+        jest.spyOn(datastore.templateSrv, 'replace').mockReturnValue('channel = "{channel1,channel2}"');
+
+        const transformQuery = datastore.transformAlarmsQueryWrapper({}, mockFilter);
+
+        expect(datastore.templateSrv.replace).toHaveBeenCalledWith('channel = "${query0}"', {});
+        expect(transformQuery).toBe('(channel = "channel1" || channel = "channel2")');
+      });
+
+      it('should transform query with multiple filters and variable combinations', () => {
+        const mockFilter = '(alarmId = "$query0" && description = "test") || channel = "Channel3"';
+        jest.spyOn(datastore.templateSrv, 'replace').mockReturnValue('(alarmId = "{alarmId1,alarmId2}" && description = "test") || channel = "Channel3"');
+
+        const transformQuery = datastore.transformAlarmsQueryWrapper({}, mockFilter);
+
+        expect(datastore.templateSrv.replace).toHaveBeenCalledWith('(alarmId = \"$query0\" && description = \"test\") || channel = \"Channel3\"', {});
+        expect(transformQuery).toBe('((alarmId = \"alarmId1\" || alarmId = \"alarmId2\") && description = \"test\") || channel = \"Channel3\"');
+      });
+
+      it('should apply the && operator for multi-value variables with the not-equals condition', () => {
+        const mockFilter = 'channel != "${query0}"';
+        jest.spyOn(datastore.templateSrv, 'replace').mockReturnValue('channel != "{channel1,channel2}"');
+
+        const transformQuery = datastore.transformAlarmsQueryWrapper({}, mockFilter);
+
+        expect(datastore.templateSrv.replace).toHaveBeenCalledWith('channel != "${query0}"', {});
+        expect(transformQuery).toBe('(channel != "channel1" && channel != "channel2")');
+      });
+
+      describe('transformSourceFilter', () => {
+        [
+          {
+            name: 'source equals',
+            input: 'source = "test-source"',
+            expected: '(properties.system = "test-source" || properties.minionId = "test-source")',
+          },
+          {
+            name: 'source does not equal',
+            input: 'source != "test-source"',
+            expected: '(properties.system != "test-source" && properties.minionId != "test-source")',
+          },
+          {
+            name: 'source is blank',
+            input: 'string.IsNullOrEmpty(source)',
+            expected: '(string.IsNullOrEmpty(properties.system) && string.IsNullOrEmpty(properties.minionId))',
+          },
+          {
+            name: 'source is not blank',
+            input: '!string.IsNullOrEmpty(source)',
+            expected: '(!string.IsNullOrEmpty(properties.system) || !string.IsNullOrEmpty(properties.minionId))',
+          },
+        ].forEach(({ name, input, expected }) => {
+          it(`should transform ${name} filter`, () => {
+            const result = datastore.transformAlarmsQueryWrapper({}, input);
+
+            expect(result).toBe(expected);
+          });
+        });
+
+        [
+          {
+            name: 'source equals',
+            input: 'source = "${query0}"',
+            replacedInput: 'source = "{source1,source2}"',
+            expected:
+              '((properties.system = "source1" || properties.system = "source2") || (properties.minionId = "source1" || properties.minionId = "source2"))',
+          },
+          {
+            name: 'source does not equal',
+            input: 'source != "${query0}"',
+            replacedInput: 'source != "{source1,source2}"',
+            expected:
+              '((properties.system != "source1" && properties.system != "source2") && (properties.minionId != "source1" && properties.minionId != "source2"))',
+          },
+        ].forEach(({ name, input, replacedInput, expected }) => {
+          it(`should transform ${name} for mutiple value variable filter`, () => {
+            jest.spyOn(datastore.templateSrv, 'replace').mockReturnValue(replacedInput);
+
+            const transformQuery = datastore.transformAlarmsQueryWrapper({}, input);
+
+            expect(datastore.templateSrv.replace).toHaveBeenCalledWith(input, {});
+            expect(transformQuery).toBe(expected);
+          });
+        });
+
+        it('should replace single value variable in the source filter', () => {
+          const mockQueryBy = 'source != "${query0}"';
+          jest.spyOn(datastore.templateSrv, 'replace').mockReturnValue('source != "test-source"');
+
+          const transformQuery = datastore.transformAlarmsQueryWrapper({}, mockQueryBy);
+
+          expect(datastore.templateSrv.replace).toHaveBeenCalledWith('source != "${query0}"', {});
+          expect(transformQuery).toBe('(properties.system != "test-source" && properties.minionId != "test-source")');
+        });
+
+        it('should handle transformation for multiple source filters in a query', () => {
+          const mockFilter = 'source = "source1" || string.IsNullOrEmpty(source)';
+
+          const result = datastore.transformAlarmsQueryWrapper({}, mockFilter);
+
+          expect(result).toBe(
+            '(properties.system = "source1" || properties.minionId = "source1") || (string.IsNullOrEmpty(properties.system) && string.IsNullOrEmpty(properties.minionId))'
+          );
+        });
+      });
     });
   });
 
@@ -171,6 +316,72 @@ describe('AlarmsQueryHandlersCore', () => {
       const result = datastore.shouldRunQuery(query);
 
       expect(result).toBe(true);
+    });
+  });
+
+  describe('loadWorkspaces', () => {
+    it('should return workspaces', async () => {
+      const workspaces = await datastore.loadWorkspaces();
+
+      expect(workspaces).toEqual(
+        new Map([
+          ['Workspace1', { id: 'Workspace1', name: 'Workspace Name' }],
+          ['Workspace2', { id: 'Workspace2', name: 'Another Workspace Name' }],
+        ])
+      );
+    });
+
+    it('should return empty map on error', async () => {
+      (datastore as any).workspaceUtils.getWorkspaces.mockRejectedValue(new Error('Error loading workspaces'));
+
+      const workspaces = await datastore.loadWorkspaces();
+
+      expect(workspaces).toEqual(new Map<string, Workspace>());
+    });
+
+    [
+      {
+        error: new Error('Request failed with status code: 404'),
+        expectedErrorDescription:
+          'The query builder lookups failed because the requested resource was not found. Please check the query parameters and try again.',
+        case: '404 error',
+      },
+      {
+        error: new Error('Request failed with status code: 429'),
+        expectedErrorDescription:
+          'The query builder lookups failed due to too many requests. Please try again later.',
+        case: '429 error',
+      },
+      {
+        error: new Error('Request failed with status code: 504'),
+        expectedErrorDescription:
+          'The query builder lookups experienced a timeout error. Some values might not be available. Narrow your query with a more specific filter and try again.',
+        case: '504 error',
+      },
+      {
+        error: new Error('Request failed with status code: 500, Error message: {"message": "Internal Server Error"}'),
+        expectedErrorDescription:
+          'Some values may not be available in the query builder lookups due to the following error: Internal Server Error.',
+        case: '500 error with message',
+      },
+      {
+        error: new Error('API failed'),
+        expectedErrorDescription:
+          'Some values may not be available in the query builder lookups due to an unknown error.',
+        case: 'Unknown error',
+      },
+    ].forEach(({ error, expectedErrorDescription, case: testCase }) => {
+      it(`should handle ${testCase}`, async () => {
+        const expectedErrorTitle = 'Warning during alarms query';
+        jest
+          .spyOn((datastore as any).workspaceUtils, 'getWorkspaces')
+          .mockRejectedValue(error);
+
+        await datastore.loadWorkspaces();
+
+        expect(datastore.errorTitle).toBe(expectedErrorTitle);
+        expect(datastore.errorDescription).toBe(expectedErrorDescription);
+      });
     });
   });
 });