+++ src/datasources/alarms/query-type-handlers/AlarmsQueryHandlersCore.ts
@@ -1,15 +1,34 @@
 import { DataSourceBase } from "core/DataSourceBase";
-import { DataQueryRequest, DataFrameDTO, TestDataSourceResponse, AppEvents, ScopedVars } from "@grafana/data";
-import { AlarmsQuery, QueryAlarmsRequest, QueryAlarmsResponse } from "../types/types";
+import { DataQueryRequest, DataFrameDTO, TestDataSourceResponse, AppEvents, ScopedVars, DataSourceInstanceSettings } from "@grafana/data";
+import { AlarmsQuery, QueryAlarmsRequest, QueryAlarmsResponse } from "./types/types";
 import { extractErrorInfo } from "core/errors";
-import { QUERY_ALARMS_RELATIVE_PATH } from "../constants/QueryAlarms.constants";
-import { ALARMS_TIME_FIELDS, AlarmsQueryBuilderFields } from "../constants/AlarmsQueryBuilder.constants";
-import { ExpressionTransformFunction, transformComputedFieldsQuery } from "core/query-builder.utils";
+import { QUERY_ALARMS_RELATIVE_PATH } from "./constants/QueryAlarms.constants";
+import { ExpressionTransformFunction, getConcatOperatorForMultiExpression, multipleValuesQuery, timeFieldsQuery, transformComputedFieldsQuery } from "core/query-builder.utils";
+import { ALARMS_TIME_FIELDS, AlarmsQueryBuilderFields } from "./constants/AlarmsQueryBuilder.constants";
+import { QueryBuilderOption, Workspace } from "core/types";
+import { WorkspaceUtils } from "shared/workspace.utils";
+import { getVariableOptions } from "core/utils";
+import { BackendSrv, getBackendSrv, getTemplateSrv, TemplateSrv } from "@grafana/runtime";
+import { MINION_ID_CUSTOM_PROPERTY, SYSTEM_CUSTOM_PROPERTY } from "./constants/SourceProperties.constants";
 
-export abstract class AlarmsQueryHandlersCore extends DataSourceBase<AlarmsQuery> {
+export abstract class AlarmsDataSourceCore extends DataSourceBase<AlarmsQuery> {
+  public errorTitle?: string;
+  public errorDescription?: string;
+
   private readonly queryAlarmsUrl = `${this.instanceSettings.url}${QUERY_ALARMS_RELATIVE_PATH}`;
+  private readonly workspaceUtils: WorkspaceUtils;
+
+  constructor(
+    readonly instanceSettings: DataSourceInstanceSettings,
+    readonly backendSrv: BackendSrv = getBackendSrv(),
+    readonly templateSrv: TemplateSrv = getTemplateSrv()
+  ) {
+    super(instanceSettings, backendSrv, templateSrv);
+    this.workspaceUtils = new WorkspaceUtils(this.instanceSettings, this.backendSrv);
+  }
 
   public abstract runQuery(query: AlarmsQuery, options: DataQueryRequest): Promise<DataFrameDTO>;
+  public readonly globalVariableOptions = (): QueryBuilderOption[] => getVariableOptions(this);
 
   protected async queryAlarms(alarmsRequestBody: QueryAlarmsRequest): Promise<QueryAlarmsResponse> {
     try {
@@ -31,6 +50,17 @@ export abstract class AlarmsQueryHandlersCore extends DataSourceBase<AlarmsQuery
     }
   }
 
+  public async loadWorkspaces(): Promise<Map<string, Workspace>> {
+    try {
+      return await this.workspaceUtils.getWorkspaces();
+    } catch (error){
+      if (!this.errorTitle) {
+        this.handleDependenciesError(error);
+      }
+      return new Map<string, Workspace>();
+    }
+  }
+
   protected transformAlarmsQuery(scopedVars: ScopedVars, query?: string): string | undefined {
     return query
       ? transformComputedFieldsQuery(this.templateSrv.replace(query, scopedVars), this.computedDataFields)
@@ -40,25 +70,52 @@ export abstract class AlarmsQueryHandlersCore extends DataSourceBase<AlarmsQuery
   private readonly computedDataFields = new Map<string, ExpressionTransformFunction>(
     Object.values(AlarmsQueryBuilderFields).map(field => {
       const dataField = field.dataField as string;
+      let callback;
 
-      return [
-        dataField,
-        this.isTimeField(dataField)
-          ? this.timeFieldsQuery(dataField)
-          : (value, operation) => `${dataField} ${operation} "${value}"`,
-      ];
+      if (dataField === AlarmsQueryBuilderFields.SOURCE.dataField) {
+        callback = this.getSourceTransformation();
+      } else if (this.isTimeField(dataField)) {
+        callback = timeFieldsQuery(dataField);
+      } else {
+        callback = multipleValuesQuery(dataField);
+      }
+
+      return [dataField, callback];
     })
   );
 
-  private timeFieldsQuery(field: string): ExpressionTransformFunction {
-    return (value: string, operation: string): string => {
-      const formattedValue = value === '${__now:date}' ? new Date().toISOString() : value;
-      return `${field} ${operation} "${formattedValue}"`;
-    };
+  private handleDependenciesError(error: unknown): void {
+    const errorDetails = extractErrorInfo((error as Error).message);
+    this.errorTitle = 'Warning during alarms query';
+    switch (errorDetails.statusCode) {
+      case '404':
+        this.errorDescription = 'The query builder lookups failed because the requested resource was not found. Please check the query parameters and try again.';
+        break;
+      case '429':
+        this.errorDescription = 'The query builder lookups failed due to too many requests. Please try again later.';
+        break;
+      case '504':
+        this.errorDescription = `The query builder lookups experienced a timeout error. Some values might not be available. Narrow your query with a more specific filter and try again.`;
+        break;
+      default:
+        this.errorDescription = errorDetails.message
+          ? `Some values may not be available in the query builder lookups due to the following error: ${errorDetails.message}.`
+          : 'Some values may not be available in the query builder lookups due to an unknown error.';
+    }
   }
 
   private isTimeField(field: string): boolean {
     return ALARMS_TIME_FIELDS.includes(field);
+  }
+
+  private getSourceTransformation(): ExpressionTransformFunction {
+    return (value: string, operation: string) => {
+      const systemExpression = multipleValuesQuery(`properties.${SYSTEM_CUSTOM_PROPERTY}`)(value, operation);
+      const minionExpression = multipleValuesQuery(`properties.${MINION_ID_CUSTOM_PROPERTY}`)(value, operation);
+      const logicalOperator = getConcatOperatorForMultiExpression(operation);
+
+      return `(${systemExpression} ${logicalOperator} ${minionExpression})`;
+    };
   }
 
   private getStatusCodeErrorMessage(errorDetails: { statusCode: string; message: string }): string {